refactor: Refactor product entities and improve formatting
This commit refactors the product entities to use `Float` instead of `Double` for price, discount, and rating, and enhances the formatting of these values across the application. Additionally, it removes unused components.

-   Updates product entities:
    -   Changes `price`, `discountPercentage`, and `rating` from `Double` to `Float` in `Product` and `ProductDetails`.
-   Refactor product response:
    -  Updates `ProductResponse` and `ProductDetailsResponse` to use `Float`.
-   Updates Product presentation:
    -   Removes `CategoryRow` and its related component `ProductItem`.
    - Updates `ProductItem` to display rating and discount percentage using the new formatting.
-   Improves value formatting:
    -   Introduces a `format` function to `Float` to handle number of decimal places.
    -   Implements platform-specific `format` functions for Android, iOS, and Desktop.
    - Updates `FilterSection` to use new formatting for price and rating.
    - Updates `HomeViewModel` to use new entity values.

+++ composeApp/src/commonMain/kotlin/org/example/cross/card/product/data/model/ProductResponse.kt
@@ -8,10 +8,10 @@ import org.example.cross.card.product.domain.entity.Product
 data class ProductResponse(
     val id: String,
     val title: String?,
-    val price: Double?,
-    val rating: Double?,
+    val price: Float?,
+    val rating: Float?,
     @SerialName("discount_percentage")
-    val discountPercentage: Double?,
+    val discountPercentage: Float?,
     @SerialName("category_id") val categoryId: String
 )
 
@@ -21,8 +21,8 @@ fun ProductResponse.toDomain(
     id = id,
     categoryId = categoryId,
     title = title ?: "Empty",
-    price = price ?: 0.0,
-    rating = rating ?: 0.0,
-    discountPercentage = discountPercentage ?: 0.0,
+    price = price ?: 0f,
+    rating = rating ?: 0f,
+    discountPercentage = discountPercentage ?: 0f,
     image = image.toDomain()
 )
